documenting props and updating examples for media object refactor to naming for cards and media vue templates
