+++ src/components/headings/typography.md
@@ -1,0 +1,54 @@
+# Typography
+
+----------
+
+Typography is the primary driver of a design’s usability and a foundational element of any interface. It’s essential for delivering a consistent, cohesive experience and should be thoughtfully managed.
+
+**Resources**
+[Presentation](https://docs.google.com/presentation/d/1fF5Auq-jtJk49_8ZvevX-_VIuE7r_uG6jNjALl9WMUA/edit?usp=sharing)
+
+
+## Typeface
+
+Sentinel is a meticulously crafted, contemporary slab serif we’ve chosen to convey our brand. It’s our primary display typeface so it’s great for page headings, titles and any big moment designed to grab a users attention. It’s also our first choice for long-form body copy. Regular, Semi-bold, and the italic variants of each the weights available for use.
+
+Roboto is an integral part of our UI. It’s open and easy-to-read letterforms are best suited to interface elements like buttons, navigation, short sections of body copy and form fields. While it comes in many weights, we’ve limited our palette to Regular, Bold and their italic counterparts.
+
+
+## Scale
+
+By using a type scale, we bring order and consistency to our communication. Our chosen ratio, a minor third, makes twelve font sizes available for use. We’ve outlined the steps taken to develop the scale:
+
+
+1. We identified our base font size of 10px. This is the starting point for the scale (but not necessarily an included size).
+  
+2. We chose a scale multiplier of 1.2 (named a minor third) and multiplied it by our base font size to yield a new value. If the result was not a whole number, we rounded it to the nearest one.
+  
+  **10 x 1.2 = 12**
+  
+3. To continue adding to the scale, we simply multiplied the previous result by the ratio again.
+  
+  **12 x 1.2 = 1.44 ≈ 14**
+
+
+## Theming
+
+We acknowledge, that over time, new typefaces will be introduced and that different digital properties require unique implementations. To anticipate future needs, we’ve built this framework with extensibility in mind.
+
+Our typographic theming convention can be thought of similarly to a slide projector. A given theme is a single slide, comprising of all the information needed to achieve a specific visual presentation. Initially, we’re rolling out with a single theme and we’re calling it Spruce. 
+
+**Spruce**
+
+- Includes Sentinel and Roboto typefaces
+- Intended for use on core products like Product Display, Expert Advice, Rich Search, and Global Navigation
+- Not intended for campaign use
+
+Looking ahead, we see additional themes being added and updates being made to existing ones. 
+
+
+## Governance
+
+**Brand.ai**
+Our design and development workflow is aided by a tool called Brand.ai. It’s the source for the most up-to-date design assets and it always reflects what’s available for designers to use. It syncs symbols and styles across our team and allows us to manage versions, edit permissions and rollout updates globally.
+
+
