+++ src/components/media/cdrMediaObject.vue
@@ -1,0 +1,108 @@
+<template>
+  <div :class="[modifierClass]">
+    <cdr-img
+    :alt="mediaFigureAlt"
+    :src="mediaFigure"
+    :ratio="ratio"
+    :crop="crop"
+    class="cdr-media-object__figure"
+    :modifier="mediaImgShape"></cdr-img>
+    <div class="cdr-media-object__body">
+      <cdr-heading :level="level">
+          <slot name="title" v-if="!mediaUrl"><span class="cdr-card__title">{{mediaTitle}}</span></slot>
+          <slot name="title" v-else><a class="cdr-card__title cdr-card__title__link" :href="mediaUrl">{{mediaTitle}}</a></slot>
+      </cdr-heading>
+      <slot v-html>{{mediaBody}}</slot>
+    </div>
+  </div>
+</template>
+
+<script>
+import cdrHeading from '../headings/cdrHeading';
+import cdrImg from '../media/cdrImg';
+import modifier from '../../mixins/modifier';
+
+export default {
+  name: 'cdr-media-object',
+  mixins: [modifier],
+  components: {
+    cdrHeading,
+    cdrImg,
+  },
+  extends: {
+    cdrImg,
+    cdrHeading,
+  },
+  props: {
+      /**
+       * Assign the proper heading level 1-6
+       **/
+    level: {
+      type: String,
+      default: '2',
+    },
+    /**
+    * Replaces the default slot with the body copy for your Media Object
+    **/
+    mediaBody: String,
+    /**
+    * Image path
+    **/
+    mediaFigure: {
+      type: String,
+      required: true,
+    },
+    /**
+    * Ratio of the media container. {square, 1-2, 2-3, 3-4, 9-16, 2-1, 3-2, 4-3, 16-9}
+    **/
+    ratio: {
+      type: String,
+      validator: value => ([
+        'square',
+        '1-2',
+        '2-3',
+        '3-4',
+        '9-16',
+        '2-1',
+        '3-2',
+        '4-3',
+        '16-9'].indexOf(value) >= 0) || false,
+    },
+    /**
+    * crop  string
+    * Area to crop the image overflow to (can be combined with ratio).
+    * {top, y-center, bottom} {left, x-center, right}
+    **/
+    crop: String,
+    /**
+    * Image alternitive description, defaults to blank
+    **/
+    mediaFigureAlt: String,
+    /**
+    * {Responsive, Rounded, circle, thumbnail}
+    **/
+    mediaImgShape: String,
+    mediaTitle: String,
+    mediaUrl: String,
+  },
+  computed: {
+    baseClass() {
+      return 'cdr-media-object';
+    },
+    ratioClass() {
+      return `cdr-media-frame--${this.ratio}`;
+    },
+    cropClass() {
+      const base = 'cdr-media-frame';
+      const cropArr = this.crop ? this.crop.split(' ') : [];
+
+      let final = '';
+
+      cropArr.forEach((crop) => {
+        final += `${base}--${crop} `;
+      });
+      return final;
+    },
+  },
+};
+</script>
